Minor breaking change with updating the datatype of `ProcFlags` from `uint` to `ulong`.

Fixes #97

+++ SimcProfileParser/Model/RawData/SimcRawSpell.cs
@@ -56,7 +56,7 @@ namespace SimcProfileParser.Model.RawData
         public uint MaxStack { get; set; }
         public uint ProcChance { get; set; }
         public int ProcCharges { get; set; }
-        public uint ProcFlags { get; set; }
+        public ulong ProcFlags { get; set; }
         public uint InternalCooldown { get; set; }
         public double Rppm { get; set; }
